unified way to interrupt worker
